requirements.txt
# requirements.txt
Flask
pandas
numpy
pulp
haversine
gunicorn
requests
googlemaps
python-dotenv